+++ cli/cli/commands/service/logs/logs.go
@@ -11,22 +11,26 @@ import (
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	docker_manager_types "github.com/kurtosis-tech/container-engine-lib/lib/docker_manager/types"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
-	positional_arg_parser "github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
+	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
 	"github.com/kurtosis-tech/kurtosis-core/commons/enclave_object_labels"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"os"
 	"strings"
 )
 
 const (
 	kurtosisLogLevelArg = "kurtosis-log-level"
-	enclaveIdArg = "enclave-id"
-	guidArg = "guid"
+	enclaveIdArg        = "enclave-id"
+	guidArg             = "guid"
+
+	shouldShowStoppedUserServiceContainers = true
+	shouldFollowContainerLogs              = false
 )
+
 var defaultKurtosisLogLevel = logrus.InfoLevel.String()
 var positionalArgs = []string{
 	enclaveIdArg,
@@ -34,10 +38,10 @@ var positionalArgs = []string{
 }
 
 var LogsCmd = &cobra.Command{
-	Use:   command_str_consts.ServiceLogsCmdStr + " [flags] " + strings.Join(positionalArgs, " "),
+	Use:                   command_str_consts.ServiceLogsCmdStr + " [flags] " + strings.Join(positionalArgs, " "),
 	DisableFlagsInUseLine: true,
-	Short: "Show logs for a service inside of an enclave",
-	RunE:  run,
+	Short:                 "Show logs for a service inside of an enclave",
+	RunE:                  run,
 }
 
 var kurtosisLogLevelStr string
@@ -81,42 +85,58 @@ func run(cmd *cobra.Command, args []string) error {
 		dockerClient,
 	)
 
-	labels := getContainerLabelsWithEnclaveIdAndGUID(enclaveId, guid)
+	labels := getUserServiceContainerLabelsWithEnclaveId(enclaveId)
 
-	containers, err := dockerManager.GetContainersByLabels(ctx, labels, true)
+	containers, err := dockerManager.GetContainersByLabels(ctx, labels, shouldShowStoppedUserServiceContainers)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting containers by labels: '%+v'", labels)
 	}
 
-	if containers != nil && len(containers) > 0 {
-		if len(containers) > 1 {
-			return stacktrace.NewError("Should exist only one container with enclave-id '%v' and guid '%v' but there are '%v' containers with these properties", enclaveId, guid, len(containers))
-		}
-
-		serviceContainer := containers[0]
-
-		readCloserLogs, err := dockerManager.GetContainerLogs(ctx, serviceContainer.GetId(), false)
-		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred getting service logs for container with ID '%v'", serviceContainer.GetId())
-		}
-		defer readCloserLogs.Close()
-
-		_, err = stdcopy.StdCopy(os.Stdout, os.Stderr, readCloserLogs)
-		if err == nil {
-			return stacktrace.Propagate(err, "An error occurred executing StdCopy")
-		}
-
+	if containers == nil || len(containers) == 0 {
+		logrus.Errorf("There is not any service container with enclave ID '%v'", enclaveId)
+		return nil
 	}
+
+	var containersWithSearchedGUID = []*docker_manager_types.Container{}
+	for _, container := range containers {
+		labelsMap := container.GetLabels()
+		containerGUID, found := labelsMap[enclave_object_labels.GUIDLabel]
+		if found && containerGUID == guid {
+			containersWithSearchedGUID = append(containersWithSearchedGUID, container)
+		}
+	}
+
+	if len(containersWithSearchedGUID) == 0 {
+		logrus.Errorf("There is not any service container with GUID '%v'", guid)
+		return nil
+	}
+
+	if len(containersWithSearchedGUID) > 1 {
+		return stacktrace.NewError("Only one container with enclave-id '%v' and GUID '%v' should exist but there are '%v' containers with these properties", enclaveId, guid, len(containers))
+	}
+
+	serviceContainer := containersWithSearchedGUID[0]
+
+	readCloserLogs, err := dockerManager.GetContainerLogs(ctx, serviceContainer.GetId(), shouldFollowContainerLogs)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting service logs for container with ID '%v'", serviceContainer.GetId())
+	}
+	defer readCloserLogs.Close()
+
+	_, err = stdcopy.StdCopy(logrus.StandardLogger().Out, logrus.StandardLogger().Out, readCloserLogs)
+	if err == nil {
+		return stacktrace.Propagate(err, "An error occurred copying the container logs to STDOUT")
+	}
+
 	return nil
 }
 
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================
-func getContainerLabelsWithEnclaveIdAndGUID(enclaveId string, guid string) map[string]string {
+func getUserServiceContainerLabelsWithEnclaveId(enclaveId string) map[string]string {
 	labels := map[string]string{}
 	labels[enclave_object_labels.ContainerTypeLabel] = enclave_object_labels.ContainerTypeUserServiceContainer
 	labels[enclave_object_labels.EnclaveIDContainerLabel] = enclaveId
-	labels[enclave_object_labels.GUIDLabel] = guid
 	return labels
 }