+++ cli/golang_internal_testsuite/networks_impl/test_network.go
@@ -6,32 +6,31 @@
 package networks_impl
 
 import (
-	"encoding/json"
-	"fmt"
-	"github.com/kurtosis-tech/example-microservice/api/api_service_client"
-	"github.com/kurtosis-tech/example-microservice/datastore/datastore_service_client"
+	"context"
+	"github.com/kurtosis-tech/example-api-server/api/golang/example_api_server_rpc_api_bindings"
+	"github.com/kurtosis-tech/example-datastore-server/api/golang/datastore_rpc_api_bindings"
+	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/client_helpers"
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/services"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"io/ioutil"
-	"os"
+	"google.golang.org/grpc"
+	"google.golang.org/protobuf/types/known/emptypb"
 	"strconv"
 )
 
 const (
-	datastoreImage                        = "kurtosistech/example-microservices_datastore"
 	datastoreServiceId services.ServiceID = "datastore"
-	datastorePort                         = 1323
 
-	apiServiceImage    = "kurtosistech/example-microservices_api"
 	apiServiceIdPrefix = "api-"
-	apiServicePort     = 2434
 
 	waitForStartupDelayMilliseconds       = 1000
 	waitForStartupMaxNumPolls             = 15
-	configFilepathRelativeToSharedDirRoot = "config-file.txt"
 )
+
+type GRPCAvailabilityChecker interface {
+	IsAvailable(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*emptypb.Empty, error)
+}
 
 type datastoreConfig struct {
 	DatastoreIp   string `json:"datastoreIp"`
@@ -41,30 +40,35 @@ type datastoreConfig struct {
 //  A custom Network implementation is intended to make test-writing easier by wrapping low-level
 //    NetworkContext calls with custom higher-level business logic
 type TestNetwork struct {
-	networkCtx                *networks.NetworkContext
-	datastoreServiceImage     string
-	apiServiceImage           string
-	datastoreClient           *datastore_service_client.DatastoreClient
-	personModifyingApiClient  *api_service_client.APIClient
-	personRetrievingApiClient *api_service_client.APIClient
-	nextApiServiceId          int
+	networkCtx                         *networks.NetworkContext
+	datastoreServiceImage              string
+	apiServiceImage                    string
+	datastoreClient                    datastore_rpc_api_bindings.DatastoreServiceClient
+	personModifyingApiClient           example_api_server_rpc_api_bindings.ExampleAPIServerServiceClient
+	personRetrievingApiClient          example_api_server_rpc_api_bindings.ExampleAPIServerServiceClient
+	personModifyingApiClientCloseFunc  func() error
+	personRetrievingApiClientCloseFunc func() error
+	nextApiServiceId                   int
 }
 
 func NewTestNetwork(networkCtx *networks.NetworkContext, datastoreServiceImage string, apiServiceImage string) *TestNetwork {
 	return &TestNetwork{
-		networkCtx:                networkCtx,
-		datastoreServiceImage:     datastoreServiceImage,
-		apiServiceImage:           apiServiceImage,
-		datastoreClient:           nil,
-		personModifyingApiClient:  nil,
-		personRetrievingApiClient: nil,
-		nextApiServiceId:          0,
+		networkCtx:                         networkCtx,
+		datastoreServiceImage:              datastoreServiceImage,
+		apiServiceImage:                    apiServiceImage,
+		datastoreClient:                    nil,
+		personModifyingApiClient:           nil,
+		personRetrievingApiClient:          nil,
+		personModifyingApiClientCloseFunc:  nil,
+		personRetrievingApiClientCloseFunc: nil,
+		nextApiServiceId:                   0,
 	}
 }
 
 //  Custom network implementations usually have a "setup" method (possibly parameterized) that is used
 //   in the Test.Setup function of each test
 func (network *TestNetwork) SetupDatastoreAndTwoApis() error {
+	ctx := context.Background()
 
 	if network.datastoreClient != nil {
 		return stacktrace.NewError("Cannot add datastore client to network; datastore client already exists!")
@@ -74,16 +78,27 @@ func (network *TestNetwork) SetupDatastoreAndTwoApis() error {
 		return stacktrace.NewError("Cannot add API services to network; one or more API services already exists")
 	}
 
-	datastoreContainerConfigSupplier := getDatastoreContainerConfigSupplier()
+	datastoreContainerConfigSupplier := client_helpers.GetDatastoreContainerConfigSupplier(network.datastoreServiceImage)
 
 	datastoreServiceContext, hostPortBindings, err := network.networkCtx.AddService(datastoreServiceId, datastoreContainerConfigSupplier)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred adding the datastore service")
 	}
 
-	datastoreClient := datastore_service_client.NewDatastoreClient(datastoreServiceContext.GetIPAddress(), datastorePort)
+	datastoreClient, datastoreClientConnCloseFunc, err := client_helpers.NewDatastoreClient(datastoreServiceContext.GetIPAddress())
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred creating a new datastore client for service with ID '%v' and IP address '%v'", datastoreServiceId, datastoreServiceContext.GetIPAddress())
+	}
+	defer func() {
+		if err := datastoreClientConnCloseFunc(); err != nil {
+			// I made this a "warn" in this case b/c there's really nothing the user can do to fix it, and it's really not a big deal if this fails
+			// In other areas where it is a big deal, and I need the user to do an action, I make it an Error with a big "ACTION REQUIRED" tag
+			//  (e.g. if our defer function to destroy a Docker network fails, so the user needs to delete the network else it'll hang around forever)
+			logrus.Warnf("We tried to close the datastore client, but doing so threw an error:\n%v", err)
+		}
+	}()
 
-	err = datastoreClient.WaitForHealthy(waitForStartupMaxNumPolls, waitForStartupDelayMilliseconds)
+	err = client_helpers.WaitForHealthy(ctx, datastoreClient, waitForStartupMaxNumPolls, waitForStartupDelayMilliseconds)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred waiting for the datastore service to become available")
 	}
@@ -92,129 +107,69 @@ func (network *TestNetwork) SetupDatastoreAndTwoApis() error {
 
 	network.datastoreClient = datastoreClient
 
-	personModifyingApiClient, err := network.addApiService()
+	personModifyingApiClient,  personModifyingApiClientCloseFunc, err := network.addApiService(ctx, datastoreServiceContext.GetIPAddress())
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred adding the person-modifying API client")
 	}
 	network.personModifyingApiClient = personModifyingApiClient
+	network.personModifyingApiClientCloseFunc = personModifyingApiClientCloseFunc
 
-	personRetrievingApiClient, err := network.addApiService()
+	personRetrievingApiClient, personRetrievingApiClientCloseFunc, err := network.addApiService(ctx, datastoreServiceContext.GetIPAddress())
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred adding the person-retrieving API client")
 	}
 	network.personRetrievingApiClient = personRetrievingApiClient
+	network.personRetrievingApiClientCloseFunc = personRetrievingApiClientCloseFunc
 
 	return nil
 }
 
 //  Custom network implementations will also usually have getters, to retrieve information about the
 //   services created during setup
-func (network *TestNetwork) GetPersonModifyingApiClient() (*api_service_client.APIClient, error) {
+func (network *TestNetwork) GetPersonModifyingApiClient() (example_api_server_rpc_api_bindings.ExampleAPIServerServiceClient, func() error, error) {
 	if network.personModifyingApiClient == nil {
-		return nil, stacktrace.NewError("No person-modifying API client exists")
+		return nil, nil, stacktrace.NewError("No person-modifying API client exists")
 	}
-	return network.personModifyingApiClient, nil
+	return network.personModifyingApiClient, network.personModifyingApiClientCloseFunc, nil
 }
-func (network *TestNetwork) GetPersonRetrievingApiClient() (*api_service_client.APIClient, error) {
+func (network *TestNetwork) GetPersonRetrievingApiClient() (example_api_server_rpc_api_bindings.ExampleAPIServerServiceClient, func() error, error) {
 	if network.personRetrievingApiClient == nil {
-		return nil, stacktrace.NewError("No person-retrieving API client exists")
+		return nil, nil, stacktrace.NewError("No person-retrieving API client exists")
 	}
-	return network.personRetrievingApiClient, nil
+	return network.personRetrievingApiClient, network.personRetrievingApiClientCloseFunc, nil
 }
 
 // ====================================================================================================
 //                                       Private helper functions
 // ====================================================================================================
-func (network *TestNetwork) addApiService() (*api_service_client.APIClient, error) {
+func (network *TestNetwork) addApiService(ctx context.Context, datastoreIp string) (example_api_server_rpc_api_bindings.ExampleAPIServerServiceClient, func() error, error) {
 
 	if network.datastoreClient == nil {
-		return nil, stacktrace.NewError("Cannot add API service to network; no datastore client exists")
+		return nil, nil, stacktrace.NewError("Cannot add API service to network; no datastore client exists")
 	}
 
 	serviceIdStr := apiServiceIdPrefix + strconv.Itoa(network.nextApiServiceId)
 	network.nextApiServiceId = network.nextApiServiceId + 1
 	serviceId := services.ServiceID(serviceIdStr)
 
-	apiServiceContainerConfigSupplier := getApiServiceContainerConfigSupplier(network)
+	apiServiceContainerConfigSupplier := client_helpers.GetApiServiceContainerConfigSupplier(network.apiServiceImage, datastoreIp)
 
 	apiServiceContext, hostPortBindings, err := network.networkCtx.AddService(serviceId, apiServiceContainerConfigSupplier)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred adding the API service")
+		return nil, nil, stacktrace.Propagate(err, "An error occurred adding the API service")
 	}
 
-	apiClient := api_service_client.NewAPIClient(apiServiceContext.GetIPAddress(), apiServicePort)
-
-	err = apiClient.WaitForHealthy(waitForStartupMaxNumPolls, waitForStartupDelayMilliseconds)
+	apiClient, apiClientCloseFunc, err := client_helpers.NewExampleAPIServerClient(apiServiceContext.GetIPAddress())
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred waiting for the api service to become available")
+		return nil, nil, stacktrace.Propagate(err, "An error occurred creating a new example API server client for service with ID '%v' and IP address '%v'", serviceId, apiServiceContext.GetIPAddress())
+	}
+
+	err = client_helpers.WaitForHealthy(ctx, apiClient, waitForStartupMaxNumPolls, waitForStartupDelayMilliseconds)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred waiting for the example API server service to become available")
 	}
 
 	logrus.Infof("Added API service with host port bindings: %+v", hostPortBindings)
-	return apiClient, nil
+	return apiClient, apiClientCloseFunc, nil
 }
 
-func getDatastoreContainerConfigSupplier() func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
-	containerConfigSupplier  := func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
-		containerConfig := services.NewContainerConfigBuilder(
-				datastoreImage,
-			).WithUsedPorts(
-				map[string]bool{fmt.Sprintf("%v/tcp", datastorePort): true},
-		    ).Build()
-		return containerConfig, nil
-	}
-	return containerConfigSupplier
-}
-
-func getApiServiceContainerConfigSupplier(network *TestNetwork) func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
-
-	containerConfigSupplier := func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
-
-		datastoreConfigFileFilePath, err := createDatastoreConfigFileInServiceDirectory(network, sharedDirectory)
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred creating data store config file in service container")
-		}
-
-		startCmd := []string{
-			"./api.bin",
-			"--config",
-			datastoreConfigFileFilePath.GetAbsPathOnServiceContainer(),
-		}
-
-		containerConfig := services.NewContainerConfigBuilder(
-				apiServiceImage,
-			).WithUsedPorts(
-				map[string]bool{fmt.Sprintf("%v/tcp", apiServicePort): true},
-			).WithCmdOverride(startCmd).Build()
-
-		return containerConfig, nil
-	}
-
-	return containerConfigSupplier
-}
-
-func createDatastoreConfigFileInServiceDirectory(network *TestNetwork, sharedDirectory *services.SharedPath) (*services.SharedPath, error) {
-	configFileFilePath := sharedDirectory.GetChildPath(configFilepathRelativeToSharedDirRoot)
-
-	logrus.Infof("Config file absolute path on this container: %v , on service container: %v", configFileFilePath.GetAbsPathOnThisContainer(), configFileFilePath.GetAbsPathOnServiceContainer())
-
-	datastoreClient := network.datastoreClient
-
-	logrus.Debugf("Datastore IP: %v , port: %v", datastoreClient.IpAddr(), datastoreClient.Port())
-
-	configObj := datastoreConfig{
-		DatastoreIp:   datastoreClient.IpAddr(),
-		DatastorePort: datastoreClient.Port(),
-	}
-	configBytes, err := json.Marshal(configObj)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred serializing the config to JSON")
-	}
-
-	logrus.Debugf("API config JSON: %v", string(configBytes))
-
-	if err := ioutil.WriteFile(configFileFilePath.GetAbsPathOnThisContainer(), configBytes, os.ModePerm); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred writing the serialized config JSON to file")
-	}
-
-	return configFileFilePath, nil
-}