+++ cli/docs/changelog.md
@@ -1,5 +1,21 @@
 # TBD
 
+# 0.6.2
+### Features
+* `enclave inspect` also prints the `Kurtosis modules`
+* Added `version_checker.CheckLatestVersion` method to check if it is running the latest CLI version before running any CLI command
+
+### Fixes
+* Fixed error where `service logs` command is executed with a nonexistent enclave ID or nonexistent GUID just silently exits
+* Upgraded to engine server 0.5.2, which returns host port bindings in the format `127.0.0.1` rather than `0.0.0.0` for Windows users
+* Upped the run timeouts of the advanced network test, module test, and bulk command execution test to 90s
+
+### Changes
+* Replaced `kurtosistech/example-microservices_datastore` Docker image with `kurtosistech/example-datastore-server` in `golang_internal_testsuite`
+* Replaced `kurtosistech/example-microservices_api` Docker image with `kurtosistech/example-api-server` in `golang_internal_testsuite`
+* Replaced `kurtosistech/example-microservices_datastore` Docker image with `docker/getting-started` in `bulk_command_execution_test` and `wait_for_endpoint_availability_test`
+* Upgraded `datastore army module` Docker image to the latest version `kurtosistech/datastore-army-module:0.1.5` in `module_test` 
+
 # 0.6.1
 ### Fixes
 * Fixed a bug where a testsuite could be reported as passing, even when the tests were failing